Fix chmod issue
If the permissions of a file were changed, umaskd would not catch this.
Added a ignore list, so that when umaskd changes the permissions, this
change does not in turn trigger another inotify event, causing an
infinite cascade.

Fixed issue with computation of new file mode. Also fixed compiler warning.
Fix automake error.